requirements.txt
appdirs==1.4.3
attrs==16.3.0
youtube_dl>=2017.8.13
pydub==0.18.0
PyQt5==5.9
PyInstaller==3.2.1
furl==1.0.0
behave==1.2.5
pyhamcrest==1.9.0
delegator.py=0.0.10
psutil==5.2.2
factory_boy==2.9.2
pypiwin32; sys_platform='win32'
